+++ src/test/java/ru/netology/stats/StatsServiceTest.java
@@ -7,10 +7,10 @@ class StatsServiceTest {
     @Test
     void shouldCalculateSum() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expected = 180;
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long expected = 180;
 
-        int actual = service.calculateSum(sales);
+        long actual = service.calculateSum(sales);
 
         assertEquals(expected, actual);
     }
@@ -19,10 +19,10 @@ class StatsServiceTest {
     @Test
     void shouldCalculateAvg() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expected = 15;
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long expected = 15;
 
-        int actual = service.calculateAvg(sales);
+        long actual = service.calculateAvg(sales);
 
         assertEquals(expected, actual);
     }
@@ -53,7 +53,7 @@ class StatsServiceTest {
     @Test
     void shouldAmountMonthUnderAvg() {
         StatsService service = new StatsService();
-        int[] amountMonthUnderAvg = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] amountMonthUnderAvg = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 5;
 
         long actual = service.amountMonthUnderAvg(amountMonthUnderAvg);
@@ -64,7 +64,7 @@ class StatsServiceTest {
     @Test
     void shouldamountMonthOverAvg() {
         StatsService service = new StatsService();
-        int[] amountMonthOverAvg = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] amountMonthOverAvg = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 5;
 
         long actual = service.amountMonthOverAvg(amountMonthOverAvg);
